% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_column.R
\name{add_column}
\alias{add_column}
\title{add_column
a function to add a column to a data frame}
\usage{
add_column(data_frame, col_names)
}
\arguments{
\item{data_frame}{data_frame to add column names}

\item{col_names}{a character vector with names of columns to add}
}
\value{
a data_frame with news columns
}
\description{
add_column
a function to add a column to a data frame
}
