% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_if_missing.R
\name{add_time_if_missing}
\alias{add_time_if_missing}
\title{add_time_if_missing
a function which add time to dates of midnight}
\usage{
add_time_if_missing(date)
}
\arguments{
\item{date}{a date}
}
\value{
a date
}
\description{
add_time_if_missing
a function which add time to dates of midnight
}
\examples{
date <- c("2024-12-03 14:50:01", "2024-12-03", "2024-12-03 14:50:24")
lapply(date, add_time_if_missing)
}
