% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date_posixct.R
\name{convert_date_posixct}
\alias{convert_date_posixct}
\title{Convert_date_posixct
a function witch allows to detected differents dates formats, used in row_data MOVE}
\usage{
convert_date_posixct(date)
}
\arguments{
\item{date}{a vecteur of date_time with the same date format}
}
\value{
a vecteur of date_time in posixct
}
\description{
Convert_date_posixct
a function witch allows to detected differents dates formats, used in row_data MOVE
}
\examples{
date <- c("03/12/24 14:50:01", "03/12/24 14:50:12", "03/12/24 14:50:24")
convert_date_posixct(date)
}
