% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_runoff.R
\name{correct_runoff}
\alias{correct_runoff}
\title{correct_runoff
a function to combined runoff event}
\usage{
correct_runoff(data, plot_name, num_event)
}
\arguments{
\item{data}{the data frame to correct}

\item{num_event}{a character vector of the id_event to combined}

\item{plot_mane}{name of the plot to correct}
}
\value{
a data_frame with combined lines
}
\description{
correct_runoff
a function to combined runoff event
}
