% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_data.R
\name{cut_data}
\alias{cut_data}
\title{cut_data
a function to cut tipping data into event}
\usage{
cut_data(data_frame, MIT)
}
\arguments{
\item{data_frame}{to data frame to cut}

\item{MIT}{a duration corresponding to minimum time inter-event}
}
\value{
the input data frame with event and a data frame of event
}
\description{
cut_data
a function to cut tipping data into event
}
