% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_10_days.R
\name{data_10_days}
\alias{data_10_days}
\title{type_rain
a function to know if rain_ev have lead to a runoff_ev}
\usage{
data_10_days(plot_name)
}
\arguments{
\item{plot}{the plot name}
}
\value{
a list with a data_frame with rain_ev to imput data and a plot of rain_ev
}
\description{
type_rain
a function to know if rain_ev have lead to a runoff_ev
}
