% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_in_days.R
\name{data_in_days}
\alias{data_in_days}
\title{data_in_days
a function to compile all data in a daily data frame}
\usage{
data_in_days(plot_name)
}
\arguments{
\item{plot_name}{the plot name}
}
\value{
a daily data_frame data
}
\description{
data_in_days
a function to compile all data in a daily data frame
}
