% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diver_correct.R
\name{diver_correct}
\alias{diver_correct}
\title{diver_correct
a function to correct diver data, with their baro_diver data}
\usage{
diver_correct(name_plot)
}
\arguments{
\item{name_plot}{name plot to correct}
}
\value{
a data_frame
}
\description{
diver_correct
a function to correct diver data, with their baro_diver data
}
