% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_runoff.R
\name{flow_runoff}
\alias{flow_runoff}
\title{flow_runoff
a function to create a data frame with the flow of runoff}
\usage{
flow_runoff(name_plot)
}
\arguments{
\item{name_plot}{a name plot to calculculate the flow}
}
\value{
a data_frame
}
\description{
flow_runoff
a function to create a data frame with the flow of runoff
}
