% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_date.R
\name{match_date}
\alias{match_date}
\title{match_date
a function to match a rain event with a runoff event}
\usage{
match_date(begin_r, date_time_rain)
}
\arguments{
\item{begin_r}{a date_time}

\item{date_time_rain}{a vector of date_time}
}
\value{
a date_time matched
}
\description{
match_date
a function to match a rain event with a runoff event
}
