% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_see.R
\name{plot_to_see}
\alias{plot_to_see}
\title{plot_to_see
a function to plot a graphs to see run off events}
\usage{
plot_to_see(name_plot, date_begin, date_end)
}
\arguments{
\item{name_plot}{name of ruisselometre to see}

\item{date_begin}{date begin of x axes}

\item{date_end}{date ed of x axes}
}
\value{
a plot
}
\description{
plot_to_see
a function to plot a graphs to see run off events
}
