% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_select.R
\name{plot_to_select}
\alias{plot_to_select}
\title{plot_to_select
a function to visual select run_off event}
\usage{
plot_to_select(name_plot)
}
\arguments{
\item{name_plot}{name of ruisselometre to display}
}
\value{
a plot
}
\description{
plot_to_select
a function to visual select run_off event
}
