% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{read_file
a function to read csv dateframe with correct features}
\usage{
read_file(read_directory)
}
\arguments{
\item{read_directory}{the file directory}
}
\value{
a dataframe
}
\description{
read_file
a function to read csv dateframe with correct features
}
