% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_file.R
\name{save_file}
\alias{save_file}
\title{save_file
a function to save a dataframe in csv with project features}
\usage{
save_file(save_directory, data_frame, date_time, name_file)
}
\arguments{
\item{save_directory}{a folder directory}

\item{data_frame}{the data frame to save}

\item{date_time}{a character vector with POSIX.CT column names}

\item{name_file}{the name of the file to save}
}
\value{
a .csv
}
\description{
save_file
a function to save a dataframe in csv with project features
}
