% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_correct.R
\name{second_correct}
\alias{second_correct}
\title{second_correct
a function to combined rain event}
\usage{
second_correct(data)
}
\arguments{
\item{data}{the data frame to correct}
}
\value{
a data_frame with news CR
}
\description{
second_correct
a function to combined rain event
}
