% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_file_path.R
\name{set_file_path}
\alias{set_file_path}
\title{Set_file_path
a function to help writting a file path}
\usage{
set_file_path(directory, folders_names)
}
\arguments{
\item{directory}{the path of package's root}

\item{folders_names}{a vector of folders names}
}
\value{
a path using folders names
}
\description{
Set_file_path
a function to help writting a file path
}
\examples{
directory <- system.file(package="moveworkflow", mustWork=TRUE)
file_names <- c("inst","data_ext","data_raw","A_1_diver")
set_file_path(directory, file_names)
}
