% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_file.R
\name{type_files}
\alias{type_files}
\title{type_files
a function which allows to set preset parameters when download rowdataframe}
\usage{
type_files(type_file)
}
\arguments{
\item{type_file}{two choice, "txt" or "CSV"}
}
\value{
a list of preset files parameters
}
\description{
type_files
a function which allows to set preset parameters when download rowdataframe
}
\examples{
file <- "txt"
type_files(file)
}
