% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_rain.R
\name{type_rain}
\alias{type_rain}
\title{type_rain
a function to know if rain_ev have lead to a runoff_ev}
\usage{
type_rain(plot)
}
\arguments{
\item{plot}{the plot name}
}
\value{
a list with a data_frame with rain_ev to imput data and a plot of rain_ev
}
\description{
type_rain
a function to know if rain_ev have lead to a runoff_ev
}
