% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_files.R
\name{combine_files}
\alias{combine_files}
\title{Combine_files
a function which combine dataframes with same setting}
\usage{
combine_files(directory, type_file)
}
\arguments{
\item{directory}{a directory with just files to combine}

\item{type_file}{type's files (ex .CSV or .txt)}
}
\value{
a dataframe of combined files
}
\description{
Combine_files
a function which combine dataframes with same setting
}
