% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_data.R
\name{cut_data}
\alias{cut_data}
\title{cut_data
a function to cut tipping data into event}
\usage{
cut_data(data_frame, MTI)
}
\arguments{
\item{data_frame}{to data frame to cut}

\item{MTI}{a duration corresponding to minimum time inter-event}
}
\value{
a raw event data_frame
}
\description{
cut_data
a function to cut tipping data into event
}
