% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rectangle.R
\name{plot_rectangle}
\alias{plot_rectangle}
\title{plot_rectangle
a function to add a rectangle to a plotly}
\usage{
plot_rectangle(data_frame, max_x)
}
\arguments{
\item{data_frame}{data_frame to add rectangle}

\item{max_x}{hight of rectangle}
}
\value{
a list to add to plotly
}
\description{
plot_rectangle
a function to add a rectangle to a plotly
}
