% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_file.R
\name{type_files}
\alias{type_files}
\title{type_files
a function which allows to preset some parameters for a dataframe download}
\usage{
type_files(type_file)
}
\arguments{
\item{type_file}{Three choices : "txt" or "CSV" or "csv"}
}
\value{
a list of preset files parameters
}
\description{
type_files
a function which allows to preset some parameters for a dataframe download
}
\examples{
file <- "txt"
type_files(file)
}
